Clone pending updates in buffered storages (#7801)
* in MmapSliceBufferedUpdateWrapper

* in MmapBitsliceBufferedUpdateWrapper

* in MutableIdTracker's versions updates

* in MutableIdTracker's mapping updates

* clone updates only when non-empty

* only lock for reconciling pending changes

* simpler reconciling

* use Mutex as argument to ensure we only lock within reconciliation

+++ lib/segment/src/id_tracker/immutable_id_tracker.rs
@@ -397,8 +397,7 @@ impl IdTracker for ImmutableIdTracker {
         );
         if has_version {
             self.internal_to_version.set(internal_id, version);
-            self.internal_to_version_wrapper
-                .set(internal_id as usize, version);
+            self.internal_to_version_wrapper.set(internal_id, version);
         }
 
         Ok(())
